refactor: extract message bubble sub-widgets into separate files
Break the 1,355-line message_bubble.dart into focused, single-responsibility
files while keeping the public API (MessageBubble + stripReplyFallback)
unchanged:

- density_metrics.dart: DensityMetrics spacing/sizing configuration
- file_bubble.dart: FileBubble for file/audio/video messages
- image_bubble.dart: ImageBubble with async encrypted image loading
- inline_reply_preview.dart: InlineReplyPreview for reply context
- hover_action_bar.dart: HoverActionBar, quick-react overlay, action icons

The main message_bubble.dart is now ~636 lines (down from 1,355). All
existing imports from consumer files remain valid with no changes needed.
