+++ lib/widgets/chat/message_bubble.dart
@@ -1,20 +1,17 @@
-import 'dart:async';
-
-import 'package:emoji_picker_flutter/emoji_picker_flutter.dart'
-    show DefaultEmojiTextStyle, EmojiPicker, Config, EmojiViewConfig,
-         CategoryViewConfig, SkinToneConfig, BottomActionBarConfig,
-         SearchViewConfig;
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:matrix/matrix.dart';
 import 'package:provider/provider.dart';
 
 import '../../services/preferences_service.dart';
-import '../../utils/media_auth.dart';
 import '../../utils/sender_color.dart';
-import '../full_image_view.dart';
 import '../user_avatar.dart';
+import 'density_metrics.dart';
+import 'file_bubble.dart';
+import 'hover_action_bar.dart';
 import 'html_message_text.dart';
+import 'image_bubble.dart';
+import 'inline_reply_preview.dart';
 import 'link_preview_card.dart';
 import 'linkable_text.dart';
 
@@ -97,7 +94,7 @@ class _MessageBubbleState extends State<MessageBubble> {
     final screenWidth = MediaQuery.sizeOf(context).width;
     final maxWidth = screenWidth * 0.72;
     final density = context.watch<PreferencesService>().messageDensity;
-    final metrics = _DensityMetrics.of(density);
+    final metrics = DensityMetrics.of(density);
     final isDesktop = screenWidth >= 720;
 
     final isRedacted = widget.event.redacted;
@@ -168,7 +165,7 @@ class _MessageBubbleState extends State<MessageBubble> {
                   children: [
                     // Hover action bar (before bubble for isMe)
                     if (isDesktop && _hovering && widget.isMe)
-                      _HoverActionBar(
+                      HoverActionBar(
                         cs: cs,
                         onReact: widget.onReact,
                         onQuickReact: widget.onQuickReact,
@@ -227,7 +224,7 @@ class _MessageBubbleState extends State<MessageBubble> {
                                     Padding(
                                       padding: const EdgeInsets.only(
                                           bottom: 4),
-                                      child: _InlineReplyPreview(
+                                      child: InlineReplyPreview(
                                         event: widget.event,
                                         timeline: widget.timeline,
                                         isMe: widget.isMe,
@@ -354,7 +351,7 @@ class _MessageBubbleState extends State<MessageBubble> {
                     ),
                     // Hover action bar (after bubble for non-me)
                     if (isDesktop && _hovering && !widget.isMe)
-                      _HoverActionBar(
+                      HoverActionBar(
                         cs: cs,
                         onReact: widget.onReact,
                         onQuickReact: widget.onQuickReact,
@@ -501,7 +498,7 @@ class _MessageBubbleState extends State<MessageBubble> {
   }
 
   Widget _buildBody(
-      BuildContext context, _DensityMetrics metrics, String bodyText) {
+      BuildContext context, DensityMetrics metrics, String bodyText) {
     final cs = Theme.of(context).colorScheme;
     final tt = Theme.of(context).textTheme;
 
@@ -533,13 +530,13 @@ class _MessageBubbleState extends State<MessageBubble> {
     }
 
     if (widget.event.messageType == MessageTypes.Image) {
-      return _ImageBubble(event: widget.event);
+      return ImageBubble(event: widget.event);
     }
 
     if (widget.event.messageType == MessageTypes.Video ||
         widget.event.messageType == MessageTypes.Audio ||
         widget.event.messageType == MessageTypes.File) {
-      return _FileBubble(
+      return FileBubble(
         event: widget.event,
         isMe: widget.isMe,
       );
@@ -636,720 +633,4 @@ String stripReplyFallback(String body) {
   // Skip the blank line after the fallback block.
   if (i < lines.length && lines[i].isEmpty) i++;
   return lines.sublist(i).join('\n');
-}
-
-// ── Inline reply preview ──────────────────────────────────────
-
-class _InlineReplyPreview extends StatefulWidget {
-  const _InlineReplyPreview({
-    required this.event,
-    required this.timeline,
-    required this.isMe,
-    this.onTap,
-  });
-
-  final Event event;
-  final Timeline? timeline;
-  final bool isMe;
-  final void Function(Event)? onTap;
-
-  @override
-  State<_InlineReplyPreview> createState() => _InlineReplyPreviewState();
-}
-
-class _InlineReplyPreviewState extends State<_InlineReplyPreview> {
-  Event? _parentEvent;
-  bool _loaded = false;
-  int _generation = 0;
-
-  @override
-  void initState() {
-    super.initState();
-    _loadParent();
-  }
-
-  @override
-  void didUpdateWidget(_InlineReplyPreview oldWidget) {
-    super.didUpdateWidget(oldWidget);
-    if (oldWidget.event != widget.event || oldWidget.timeline != widget.timeline) {
-      _generation++;
-      _loaded = false;
-      _parentEvent = null;
-      _loadParent();
-    }
-  }
-
-  Future<void> _loadParent() async {
-    final gen = _generation;
-    if (widget.timeline == null) {
-      if (mounted && gen == _generation) setState(() => _loaded = true);
-      return;
-    }
-    try {
-      final parent = await widget.event.getReplyEvent(widget.timeline!);
-      if (mounted && gen == _generation) {
-        setState(() {
-          _parentEvent = parent;
-          _loaded = true;
-        });
-      }
-    } catch (e) {
-      debugPrint('[Lattice] Failed to load reply parent: $e');
-      if (mounted && gen == _generation) setState(() => _loaded = true);
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    if (!_loaded) return const SizedBox.shrink();
-
-    final cs = Theme.of(context).colorScheme;
-    final tt = Theme.of(context).textTheme;
-
-    final parentAvailable =
-        _parentEvent != null &&
-        _parentEvent!.type != EventTypes.Redaction &&
-        !_parentEvent!.redacted;
-    final senderName = parentAvailable
-        ? (_parentEvent!.senderFromMemoryOrFallback.displayName ??
-            _parentEvent!.senderId)
-        : null;
-    final color = parentAvailable
-        ? senderColor(_parentEvent!.senderId, cs)
-        : cs.onSurfaceVariant;
-
-    return GestureDetector(
-      onTap: parentAvailable ? () => widget.onTap?.call(_parentEvent!) : null,
-      child: Container(
-        constraints: const BoxConstraints(minHeight: 36),
-        padding: const EdgeInsets.only(left: 8, right: 4, top: 4, bottom: 4),
-        decoration: BoxDecoration(
-          border: Border(left: BorderSide(color: color, width: 2)),
-          color: widget.isMe
-              ? cs.onPrimary.withValues(alpha: 0.12)
-              : cs.onSurface.withValues(alpha: 0.06),
-          borderRadius: const BorderRadius.only(
-            topRight: Radius.circular(4),
-            bottomRight: Radius.circular(4),
-          ),
-        ),
-        child: parentAvailable
-            ? Column(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: [
-                  Text(
-                    senderName!,
-                    maxLines: 1,
-                    overflow: TextOverflow.ellipsis,
-                    style: tt.bodySmall?.copyWith(
-                      fontWeight: FontWeight.w600,
-                      color: color,
-                    ),
-                  ),
-                  Text(
-                    stripReplyFallback(_parentEvent!.body),
-                    maxLines: 1,
-                    overflow: TextOverflow.ellipsis,
-                    style: tt.bodySmall?.copyWith(
-                      color: widget.isMe
-                          ? cs.onPrimary.withValues(alpha: 0.7)
-                          : cs.onSurfaceVariant,
-                    ),
-                  ),
-                ],
-              )
-            : Text(
-                'Message not available',
-                style: tt.bodySmall?.copyWith(
-                  fontStyle: FontStyle.italic,
-                  color: widget.isMe
-                      ? cs.onPrimary.withValues(alpha: 0.5)
-                      : cs.onSurfaceVariant.withValues(alpha: 0.5),
-                ),
-              ),
-      ),
-    );
-  }
-}
-
-// ── Hover reply button ────────────────────────────────────────
-
-class _HoverActionBar extends StatefulWidget {
-  const _HoverActionBar({
-    required this.cs,
-    this.onReact,
-    this.onQuickReact,
-    this.onReply,
-    required this.onMore,
-    this.onQuickReactOpenChanged,
-  });
-
-  final ColorScheme cs;
-  final VoidCallback? onReact;
-  final void Function(String emoji)? onQuickReact;
-  final VoidCallback? onReply;
-  final void Function(Offset position) onMore;
-
-  /// Notifies the parent when the quick-react overlay opens/closes.
-  final ValueChanged<bool>? onQuickReactOpenChanged;
-
-  @override
-  State<_HoverActionBar> createState() => _HoverActionBarState();
-}
-
-class _HoverActionBarState extends State<_HoverActionBar> {
-  OverlayEntry? _overlayEntry;
-  final LayerLink _layerLink = LayerLink();
-
-  @override
-  void dispose() {
-    _removeOverlay();
-    super.dispose();
-  }
-
-  void _removeOverlay() {
-    _overlayEntry?.remove();
-    _overlayEntry = null;
-    widget.onQuickReactOpenChanged?.call(false);
-  }
-
-  void _showQuickReactPopup() {
-    final overlay = Overlay.of(context);
-    final box = context.findRenderObject() as RenderBox?;
-    if (box == null || !box.hasSize) return;
-
-    widget.onQuickReactOpenChanged?.call(true);
-
-    _overlayEntry = OverlayEntry(
-      builder: (ctx) => _QuickReactOverlay(
-        link: _layerLink,
-        anchorSize: box.size,
-        cs: widget.cs,
-        hasMore: widget.onReact != null,
-        onEmojiSelected: (emoji) {
-          _removeOverlay();
-          widget.onQuickReact?.call(emoji);
-        },
-        onDismiss: _removeOverlay,
-      ),
-    );
-    overlay.insert(_overlayEntry!);
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final hasReact = widget.onReact != null || widget.onQuickReact != null;
-    return CompositedTransformTarget(
-      link: _layerLink,
-      child: Padding(
-        padding: const EdgeInsets.symmetric(horizontal: 4),
-        child: Material(
-          elevation: 2,
-          borderRadius: BorderRadius.circular(16),
-          color: widget.cs.surfaceContainerHighest,
-          child: Row(
-            mainAxisSize: MainAxisSize.min,
-            children: [
-              if (hasReact)
-                _ActionIcon(
-                  icon: Icons.add_reaction_outlined,
-                  onTap: _showQuickReactPopup,
-                  cs: widget.cs,
-                  borderRadius: const BorderRadius.horizontal(
-                    left: Radius.circular(16),
-                  ),
-                ),
-              if (widget.onReply != null)
-                _ActionIcon(
-                  icon: Icons.reply_rounded,
-                  onTap: widget.onReply!,
-                  cs: widget.cs,
-                ),
-              _ActionIcon(
-                icon: Icons.more_horiz_rounded,
-                onTap: () {
-                  final box = context.findRenderObject() as RenderBox?;
-                  if (box == null || !box.hasSize) return;
-                  final pos = box.localToGlobal(
-                    Offset(box.size.width, box.size.height / 2),
-                  );
-                  widget.onMore(pos);
-                },
-                cs: widget.cs,
-                borderRadius: const BorderRadius.horizontal(
-                  right: Radius.circular(16),
-                ),
-              ),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class _ActionIcon extends StatelessWidget {
-  const _ActionIcon({
-    required this.icon,
-    required this.onTap,
-    required this.cs,
-    this.borderRadius,
-  });
-
-  final IconData icon;
-  final VoidCallback onTap;
-  final ColorScheme cs;
-  final BorderRadius? borderRadius;
-
-  @override
-  Widget build(BuildContext context) {
-    return InkWell(
-      borderRadius: borderRadius ?? BorderRadius.zero,
-      onTap: onTap,
-      child: Padding(
-        padding: const EdgeInsets.all(6),
-        child: Icon(icon, size: 16, color: cs.onSurfaceVariant),
-      ),
-    );
-  }
-}
-
-// ── Quick-react overlay ─────────────────────────────────────
-
-class _QuickReactOverlay extends StatefulWidget {
-  const _QuickReactOverlay({
-    required this.link,
-    required this.anchorSize,
-    required this.cs,
-    required this.hasMore,
-    required this.onEmojiSelected,
-    required this.onDismiss,
-  });
-
-  final LayerLink link;
-  final Size anchorSize;
-  final ColorScheme cs;
-
-  /// Whether to show the "..." button (only when a full emoji picker is available).
-  final bool hasMore;
-  final void Function(String emoji) onEmojiSelected;
-  final VoidCallback onDismiss;
-
-  @override
-  State<_QuickReactOverlay> createState() => _QuickReactOverlayState();
-}
-
-class _QuickReactOverlayState extends State<_QuickReactOverlay> {
-  bool _showPicker = false;
-
-  static const _quickEmojis = [
-    '\u{2764}\u{FE0F}', // ❤️
-    '\u{1F44D}', // 👍
-    '\u{1F44E}', // 👎
-    '\u{1F602}', // 😂
-    '\u{1F622}', // 😢
-    '\u{1F62E}', // 😮
-  ];
-
-  @override
-  Widget build(BuildContext context) {
-    final cs = widget.cs;
-    const gap = 4.0;
-
-    return Stack(
-      children: [
-        // Dismiss scrim
-        Positioned.fill(
-          child: GestureDetector(
-            behavior: HitTestBehavior.translucent,
-            onTap: widget.onDismiss,
-          ),
-        ),
-        CompositedTransformFollower(
-          link: widget.link,
-          targetAnchor: Alignment.topCenter,
-          followerAnchor: Alignment.bottomCenter,
-          offset: const Offset(0, -gap),
-          child: UnconstrainedBox(
-            child: Column(
-              mainAxisSize: MainAxisSize.min,
-              children: [
-                // Full emoji picker (above the quick-react bar)
-                if (_showPicker)
-                  Padding(
-                    padding: const EdgeInsets.only(bottom: gap),
-                    child: Material(
-                      elevation: 4,
-                      borderRadius: BorderRadius.circular(16),
-                      color: cs.surfaceContainer,
-                      clipBehavior: Clip.antiAlias,
-                      child: SizedBox(
-                        width: 350,
-                        height: 400,
-                        child: EmojiPicker(
-                          onEmojiSelected: (category, emoji) {
-                            widget.onEmojiSelected(emoji.emoji);
-                          },
-                          config: Config(
-                            emojiTextStyle: DefaultEmojiTextStyle,
-                            emojiViewConfig: EmojiViewConfig(
-                              columns: 8,
-                              emojiSizeMax: 28,
-                              backgroundColor: cs.surfaceContainer,
-                            ),
-                            categoryViewConfig: CategoryViewConfig(
-                              backgroundColor: cs.surfaceContainer,
-                              indicatorColor: cs.primary,
-                              iconColorSelected: cs.primary,
-                              iconColor: cs.onSurfaceVariant,
-                            ),
-                            skinToneConfig: SkinToneConfig(
-                              dialogBackgroundColor:
-                                  cs.surfaceContainerHighest,
-                              indicatorColor: cs.primary,
-                            ),
-                            bottomActionBarConfig: BottomActionBarConfig(
-                              backgroundColor: cs.surfaceContainer,
-                              buttonColor: cs.primary,
-                            ),
-                            searchViewConfig: SearchViewConfig(
-                              backgroundColor: cs.surfaceContainer,
-                              buttonIconColor: cs.onSurfaceVariant,
-                            ),
-                          ),
-                        ),
-                      ),
-                    ),
-                  ),
-                // Quick-react bar
-                Material(
-                  elevation: 4,
-                  borderRadius: BorderRadius.circular(20),
-                  color: cs.surfaceContainerHighest,
-                  clipBehavior: Clip.antiAlias,
-                  child: Padding(
-                    padding: const EdgeInsets.symmetric(horizontal: 4),
-                    child: Row(
-                      mainAxisSize: MainAxisSize.min,
-                      children: [
-                        for (final emoji in _quickEmojis)
-                          InkWell(
-                            borderRadius: BorderRadius.circular(20),
-                            onTap: () => widget.onEmojiSelected(emoji),
-                            child: Padding(
-                              padding: const EdgeInsets.all(6),
-                              child: Text(
-                                emoji,
-                                style: DefaultEmojiTextStyle.copyWith(
-                                    fontSize: 22),
-                              ),
-                            ),
-                          ),
-                        if (widget.hasMore)
-                          InkWell(
-                            borderRadius: BorderRadius.circular(20),
-                            onTap: () =>
-                                setState(() => _showPicker = !_showPicker),
-                            child: Padding(
-                              padding: const EdgeInsets.all(6),
-                              child: Icon(
-                                Icons.more_horiz_rounded,
-                                size: 22,
-                                color: cs.onSurfaceVariant,
-                              ),
-                            ),
-                          ),
-                      ],
-                    ),
-                  ),
-                ),
-              ],
-            ),
-          ),
-        ),
-      ],
-    );
-  }
-}
-
-// ── Density metrics ──────────────────────────────────────────
-
-class _DensityMetrics {
-  const _DensityMetrics({
-    required this.firstMessageTopPad,
-    required this.messageTopPad,
-    required this.messageBottomPad,
-    required this.avatarRadius,
-    required this.avatarFontSize,
-    required this.bubbleHorizontalPad,
-    required this.bubbleVerticalPad,
-    required this.bubbleRadius,
-    required this.senderNameBottomPad,
-    required this.senderNameFontSize,
-    required this.bodyFontSize,
-    required this.bodyLineHeight,
-    required this.timestampTopPad,
-    required this.timestampFontSize,
-    required this.statusIconSize,
-  });
-
-  final double firstMessageTopPad;
-  final double messageTopPad;
-  final double messageBottomPad;
-  final double avatarRadius;
-  final double avatarFontSize;
-  final double bubbleHorizontalPad;
-  final double bubbleVerticalPad;
-  final double bubbleRadius;
-  final double senderNameBottomPad;
-  final double senderNameFontSize;
-  final double bodyFontSize;
-  final double bodyLineHeight;
-  final double timestampTopPad;
-  final double timestampFontSize;
-  final double statusIconSize;
-
-  static const _compact = _DensityMetrics(
-    firstMessageTopPad: 6,
-    messageTopPad: 1,
-    messageBottomPad: 1,
-    avatarRadius: 12,
-    avatarFontSize: 9,
-    bubbleHorizontalPad: 10,
-    bubbleVerticalPad: 6,
-    bubbleRadius: 16,
-    senderNameBottomPad: 2,
-    senderNameFontSize: 11,
-    bodyFontSize: 13,
-    bodyLineHeight: 1.3,
-    timestampTopPad: 3,
-    timestampFontSize: 9,
-    statusIconSize: 12,
-  );
-
-  static const _default = _DensityMetrics(
-    firstMessageTopPad: 10,
-    messageTopPad: 2,
-    messageBottomPad: 2,
-    avatarRadius: 14,
-    avatarFontSize: 11,
-    bubbleHorizontalPad: 14,
-    bubbleVerticalPad: 9,
-    bubbleRadius: 18,
-    senderNameBottomPad: 3,
-    senderNameFontSize: 12,
-    bodyFontSize: 14,
-    bodyLineHeight: 1.4,
-    timestampTopPad: 4,
-    timestampFontSize: 10,
-    statusIconSize: 13,
-  );
-
-  static const _comfortable = _DensityMetrics(
-    firstMessageTopPad: 14,
-    messageTopPad: 4,
-    messageBottomPad: 4,
-    avatarRadius: 16,
-    avatarFontSize: 12,
-    bubbleHorizontalPad: 16,
-    bubbleVerticalPad: 12,
-    bubbleRadius: 20,
-    senderNameBottomPad: 4,
-    senderNameFontSize: 13,
-    bodyFontSize: 15,
-    bodyLineHeight: 1.5,
-    timestampTopPad: 5,
-    timestampFontSize: 11,
-    statusIconSize: 14,
-  );
-
-  static _DensityMetrics of(MessageDensity density) => switch (density) {
-        MessageDensity.compact => _compact,
-        MessageDensity.defaultDensity => _default,
-        MessageDensity.comfortable => _comfortable,
-      };
-
-}
-
-// ── File bubble (video / audio / generic file) ───────────────
-
-class _FileBubble extends StatelessWidget {
-  const _FileBubble({required this.event, required this.isMe});
-
-  final Event event;
-  final bool isMe;
-
-  @override
-  Widget build(BuildContext context) {
-    final cs = Theme.of(context).colorScheme;
-    final tt = Theme.of(context).textTheme;
-    final foreground = isMe ? cs.onPrimary : cs.onSurface;
-
-    final icon = switch (event.messageType) {
-      MessageTypes.Video => Icons.videocam_rounded,
-      MessageTypes.Audio => Icons.audiotrack_rounded,
-      _ => Icons.insert_drive_file_rounded,
-    };
-
-    final fileName = event.body;
-    final infoMap = event.content.tryGet<Map<String, Object?>>('info');
-    final fileSize = infoMap?.tryGet<int>('size');
-
-    return Container(
-      padding: const EdgeInsets.symmetric(horizontal: 4, vertical: 4),
-      child: Row(
-        mainAxisSize: MainAxisSize.min,
-        children: [
-          Icon(icon, size: 28, color: foreground.withValues(alpha: 0.7)),
-          const SizedBox(width: 8),
-          Flexible(
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                Text(
-                  fileName,
-                  maxLines: 2,
-                  overflow: TextOverflow.ellipsis,
-                  style: tt.bodyMedium?.copyWith(
-                    color: foreground,
-                    fontWeight: FontWeight.w500,
-                  ),
-                ),
-                if (fileSize != null)
-                  Text(
-                    _formatFileSize(fileSize),
-                    style: tt.bodySmall?.copyWith(
-                      color: foreground.withValues(alpha: 0.6),
-                    ),
-                  ),
-              ],
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-
-  static String _formatFileSize(int bytes) {
-    if (bytes < 1024) return '$bytes B';
-    if (bytes < 1024 * 1024) return '${(bytes / 1024).toStringAsFixed(1)} KB';
-    if (bytes < 1024 * 1024 * 1024) {
-      return '${(bytes / (1024 * 1024)).toStringAsFixed(1)} MB';
-    }
-    return '${(bytes / (1024 * 1024 * 1024)).toStringAsFixed(1)} GB';
-  }
-}
-
-// ── Image bubble (async URI resolution) ──────────────────────
-
-class _ImageBubble extends StatefulWidget {
-  const _ImageBubble({required this.event});
-
-  final Event event;
-
-  @override
-  State<_ImageBubble> createState() => _ImageBubbleState();
-}
-
-class _ImageBubbleState extends State<_ImageBubble> {
-  Uint8List? _imageBytes;
-  String? _imageUrl;
-  bool _loading = true;
-
-  @override
-  void initState() {
-    super.initState();
-    _loadImage();
-  }
-
-  @override
-  void didUpdateWidget(_ImageBubble old) {
-    super.didUpdateWidget(old);
-    if (old.event.eventId != widget.event.eventId) {
-      _imageBytes = null;
-      _imageUrl = null;
-      _loading = true;
-      _loadImage();
-    }
-  }
-
-  Future<void> _loadImage() async {
-    try {
-      if (widget.event.isAttachmentEncrypted) {
-        final file = await widget.event.downloadAndDecryptAttachment(
-          getThumbnail: true,
-        );
-        if (mounted) {
-          setState(() {
-            _imageBytes = file.bytes;
-            _loading = false;
-          });
-        }
-      } else {
-        final uri = await widget.event.getAttachmentUri(
-          getThumbnail: true,
-          width: 280,
-          height: 260,
-          method: ThumbnailMethod.scale,
-        );
-        if (mounted) {
-          setState(() {
-            _imageUrl = uri?.toString();
-            _loading = false;
-          });
-        }
-      }
-    } catch (e) {
-      debugPrint('[Lattice] Image bubble load failed: $e');
-      if (mounted) setState(() => _loading = false);
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final cs = Theme.of(context).colorScheme;
-
-    return GestureDetector(
-      onTap: () => showFullImageDialog(context, widget.event),
-      child: ClipRRect(
-        borderRadius: BorderRadius.circular(10),
-        child: ConstrainedBox(
-          constraints: const BoxConstraints(maxHeight: 260, maxWidth: 280),
-          child: _loading
-              ? Container(
-                  height: 80,
-                  color: cs.surfaceContainerHighest,
-                  child: const Center(
-                    child: CircularProgressIndicator(strokeWidth: 2),
-                  ),
-                )
-              : _imageBytes != null
-                  ? Image.memory(_imageBytes!, fit: BoxFit.cover)
-                  : _imageUrl != null
-                      ? Image.network(
-                          _imageUrl!,
-                          fit: BoxFit.cover,
-                          headers: mediaAuthHeaders(
-                            widget.event.room.client,
-                            _imageUrl!,
-                          ),
-                          errorBuilder: (_, __, ___) => Container(
-                            height: 80,
-                            color: cs.surfaceContainerHighest,
-                            child: const Center(
-                              child: Icon(Icons.broken_image_outlined),
-                            ),
-                          ),
-                        )
-                      : Container(
-                          height: 80,
-                          color: cs.surfaceContainerHighest,
-                          child: const Center(
-                            child: Icon(Icons.broken_image_outlined),
-                          ),
-                        ),
-        ),
-      ),
-    );
-  }
 }
